Changed all char to byte - unicode did not work well with networking

+++ src/userInputHandling/KeyHandler.java
@@ -55,7 +55,7 @@ public class KeyHandler implements KeyListener {
 		
 	  public synchronized void keyPressed( KeyEvent e ) {
 		  System.out.println("Local key pressed: " + e.getKeyChar());
-		  networking.Network.nextKey = e.getKeyChar();
+		  networking.Network.nextKey = (byte) e.getKeyChar();
 		  
 		  	/*int keyCode = e.getKeyCode();
 	    	if( keyCode >= 0 && keyCode < KEY_COUNT ) {
